Refactor "is a potentially valid key range"
