+++ src/lib/isPotentiallyValidKeyRange.ts
@@ -1,32 +1,21 @@
 import FDBKeyRange from "../FDBKeyRange.js";
-import isSharedArrayBuffer from "./isSharedArrayBuffer.js";
+import valueToKeyWithoutThrowing, {
+    INVALID_TYPE,
+} from "./valueToKeyWithoutThrowing.js";
 
 // https://www.w3.org/TR/IndexedDB/#is-a-potentially-valid-key-range
 const isPotentiallyValidKeyRange = (value: any): boolean => {
-    // Many of these conditions recapitulate the same conditions in `valueToKeyRange.ts`
-    return (
-        // FDBKeyRange
-        value instanceof FDBKeyRange ||
-        // null/undefined - nullDisallowedFlag=false is assumed here
-        value === null ||
-        value === undefined ||
-        // Number
-        typeof value === "number" ||
-        // Date
-        Object.prototype.toString.call(value) === "[object Date]" ||
-        // string
-        typeof value === "string" ||
-        // buffer source type
-        // note: we are explicitly _not_ checking for detachedness here, to match Chromium's behavior
-        // see: https://github.com/w3c/IndexedDB/issues/465
-        value instanceof ArrayBuffer ||
-        isSharedArrayBuffer(value) ||
-        (typeof ArrayBuffer !== "undefined" &&
-            ArrayBuffer.isView &&
-            ArrayBuffer.isView(value)) ||
-        // array exotic type
-        Array.isArray(value)
-    );
+    // If value is a key range, return true.
+    if (value instanceof FDBKeyRange) {
+        return true;
+    }
+
+    // Let key be the result of converting a value to a key with value.
+    const key = valueToKeyWithoutThrowing(value);
+
+    // If key is "invalid type" return false.
+    // Else return true.
+    return key !== INVALID_TYPE;
 };
 
 export default isPotentiallyValidKeyRange;
